+++ libs/sdk-core/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `getLabelFromFilter` and `getFilterFromLabel` functions
 - Add NUA key to the local storage only when the key is created from the desktop application
 - Update `/processing/status` endpoint response model
+- Update `getParagraphText` and `getSentenceText` methods to improve performance
 
 # 1.0.6 (2022-11-17)